requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
matplotlib>=3.7.0
Pillow>=9.5.0
opencv-python>=4.8.0
torchview>=0.1.2
transformers>=4.40.0
